refactor: 사용자 관련 수정
